Glitch: The soundcloud 'track progress' bar is in front of the play button, so a user trying to press pause would find that pressing the lower half of the pause button would only rewind the track to the first few seconds, to matter how ofter they pressed it. Annoying!
Fix: made the soundcloud widget a little taller, fixing the conflict.

Remaining: the soundcloud widget takes too much space! So does the New Thread form. The New Thread form should hover in the bottom right as the user scrolls down instead of being stuck at the top of the page.